rename: in -> input , out -> output 으로 수정
