+++ src/main/java/subway/application/service/LineLoadService.java
@@ -2,7 +2,7 @@ package subway.application.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.application.service.out.LineRepositoryPort;
+import subway.application.service.output.LineRepositoryPort;
 import subway.domain.LineDomain;
 import subway.domain.NotFoundLineException;
